Move constant filenames from job_runner to ert
The import of these filenames are done also from res,
and res importing job_runner causes a circular import.
