+++ src/ert/constant_filenames.py
@@ -1,0 +1,11 @@
+"""
+Contains constant definitions of file names used by
+ert.
+"""
+CERT_FILE = ".ee.pem"
+JOBS_FILE = "jobs.json"
+LOG_file = "JOB_LOG"
+ERROR_file = "ERROR"
+STATUS_file = "STATUS"
+OK_file = "OK"
+STATUS_json = "status.json"
