--- src/ert/job_runner/__init__.py
@@ -1,7 +1,0 @@
-CERT_FILE = ".ee.pem"
-JOBS_FILE = "jobs.json"
-LOG_file = "JOB_LOG"
-ERROR_file = "ERROR"
-STATUS_file = "STATUS"
-OK_file = "OK"
-STATUS_json = "status.json"
